requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
pillow
shap
